# shared-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/properties/{property_id}": {
      "get": {
        "tags": [
          "property"
        ],
        "summary": "Get property by ID (FR-PROP-02)",
        "description": "Retrieve a single property by its UUID.\n\nReturns 404 if the property does not exist.",
        "operationId": "get_property_api_v1_properties__property_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties": {
      "get": {
        "tags": [
          "property"
        ],
        "summary": "List all properties",
        "description": "GET /api/v1/properties \u2014 list all properties.",
        "operationId": "list_properties_api_v1_properties_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "property"
        ],
        "summary": "Create a new property (FR-PROP-01)",
        "description": "Create a new property.\n\nThe caller becomes the owner of the property.\nAudit log is recorded automatically by the service.",
        "operationId": "create_property_api_v1_properties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/properties/{property_id}/rooms": {
      "get": {
        "tags": [
          "property"
        ],
        "summary": "Get property with all its rooms (FR-PROP-06)",
        "description": "Retrieve a property with its full room list.",
        "operationId": "get_property_rooms_api_v1_properties__property_id__rooms_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyWithRoomsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/{room_id}/status": {
      "patch": {
        "tags": [
          "property"
        ],
        "summary": "Update room status (FR-PROP-05)",
        "description": "Update a room's current status (available/occupied/maintenance).\n\nAudit log is recorded automatically by the service.",
        "operationId": "update_room_status_api_v1_rooms__room_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Room Status Api V1 Rooms  Room Id  Status Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Send a test notification",
        "description": "Creates and sends a test notification (mock). Fail-silent \u2014 always returns 201.",
        "operationId": "send_test_notification_api_v1_notifications_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/history": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get notification history",
        "description": "GET /api/v1/notifications/history.",
        "operationId": "get_notification_history_api_v1_notifications_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by property",
              "title": "Property Id"
            },
            "description": "Filter by property"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notif_id}/resend": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "Resend a failed notification",
        "description": "PATCH /api/v1/notifications/{id}/resend.",
        "operationId": "resend_notification_api_v1_notifications__notif_id__resend_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notif_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notif Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePropertyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "title": "Address"
          },
          "billing_due_day": {
            "type": "integer",
            "maximum": 28.0,
            "minimum": 1.0,
            "title": "Billing Due Day"
          },
          "min_deposit_months": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Min Deposit Months",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "address",
          "billing_due_day"
        ],
        "title": "CreatePropertyRequest",
        "description": "Request body for POST /api/v1/properties.\n\nValidation:\n    - ``billing_due_day`` must be 1-28.\n    - ``min_deposit_months`` >= 1.\n    - ``name`` must not be empty."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationChannel": {
        "type": "string",
        "enum": [
          "email",
          "line",
          "sms"
        ],
        "title": "NotificationChannel"
      },
      "NotificationCreateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationResponse"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "NotificationCreateResponse"
      },
      "NotificationListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "NotificationListResponse"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "channel",
          "subject",
          "body",
          "status",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
...
      "PropertyCreateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PropertyResponse"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "PropertyCreateResponse",
        "description": "Wrapper for 201 CREATED response (SDD \u00a73.1)."
      },
      "PropertyListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PropertyResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "PropertyListResponse",
        "description": "Response for GET /api/v1/properties \u2014 list all properties."
      },
      "PropertyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "billing_due_day": {
            "type": "integer",
            "title": "Billing Due Day"
          },
          "min_deposit_months": {
            "type": "integer",
            "title": "Min Deposit Months"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "billing_due_day",
          "min_deposit_months",
          "created_at",
          "updated_at"
        ],
        "title": "PropertyResponse",
        "description": "Response body for a single Property resource (SDD \u00a73.1 format)."
      },
      "PropertyWithRoomsResponse": {
        "properties": {
          "property": {
            "$ref": "#/components/schemas/PropertyResponse"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/RoomResponse"
            },
            "type": "array",
            "title": "Rooms"
          }
        },
        "type": "object",
        "required": [
          "property",
          "rooms"
        ],
        "title": "PropertyWithRoomsResponse",
        "description": "Response for GET /api/v1/properties/{id}/rooms."
      },
...
      "RoomResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "building_id": {
            "type": "string",
            "format": "uuid",
            "title": "Building Id"
          },
          "floor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor Id"
          },
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "base_rent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Base Rent"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "images": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "building_id",
          "room_number",
          "room_type",
          "base_rent",
          "status"
        ],
        "title": "RoomResponse",
        "description": "Response body for a single Room resource (SDD \u00a73.1 format).\n\nNote: ``images`` is omitted from the response if ``None``."
      },
      "RoomStatus": {
        "type": "string",
        "enum": [
          "available",
          "occupied",
          "maintenance"
        ],
        "title": "RoomStatus",
        "description": "Possible states for a room (SDD \u00a76: State Machines)."
      },
      "SendNotificationRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "channel": {
            "$ref": "#/components/schemas/NotificationChannel",
            "default": "email"
          },
          "subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "property_id",
          "subject",
          "body"
        ],
        "title": "SendNotificationRequest"
      },
...
      "UpdateRoomStatusRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RoomStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UpdateRoomStatusRequest",
        "description": "Request body for PATCH /api/v1/rooms/{id}/status."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}